% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{auditInitiate}
\alias{auditInitiate}
\title{auditInitiate}
\usage{
auditInitiate(conn, idUser, idSession, idProject, idResult, researchGoal,
  exportType)
}
\arguments{
\item{conn}{connection object}
}
\description{
From the result of the query i2b2, this function returns a list of matching patients.
}

