% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{auditLogError}
\alias{auditLogError}
\title{auditLogError}
\usage{
auditLogError(conn, rplugin_audit_id, error)
}
\arguments{
\item{conn}{connection object}
}
\description{
From the result of the query i2b2, this function returns a list of matching patients.
}

