% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{auditLogOk}
\alias{auditLogOk}
\title{auditLogOk}
\usage{
auditLogOk(conn, rplugin_audit_id)
}
\arguments{
\item{conn}{connection object}
}
\description{
From the result of the query i2b2, this function returns a list of matching patients.
}

