% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{auditLogRole}
\alias{auditLogRole}
\title{auditLogRole}
\usage{
auditLogRole(conn, rplugin_audit_id, role)
}
\description{
From the result of the query i2b2, this function returns a list of matching patients.
}

