% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{buildColumn}
\alias{buildColumn}
\title{buildColumn}
\usage{
buildColumn(column, prefix)
}
\arguments{
\item{character}{the column}

\item{character}{the prefix table}
}
\value{
character List of columns to export
}
\description{
builds columns to display for exporting visit data set
}

