% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{conceptDimension}
\alias{conceptDimension}
\title{conceptDimension}
\usage{
conceptDimension(conn)
}
\arguments{
\item{conn}{connection object}
}
\value{
data.frame results
}
\description{
From the result of the query i2b2, this function returns a list of matching patients.
}

