% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{downloadAsCsv}
\alias{downloadAsCsv}
\title{downloadAsCsv}
\usage{
downloadAsCsv(idUser, idResult, idProject, idSession, researchGoal, exportType)
}
\arguments{
\item{character}{idUser User ID}

\item{numeric}{idResult Result number}

\item{character}{idProject Projet ID}

\item{character}{idSession Session ID}

\item{character}{researchGoal Study goal}

\item{character}{exportData data type to export}
}
\description{
This function creates files by each type of data to export.
}

