% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{encounterCollection}
\alias{encounterCollection}
\title{encounterCollection}
\usage{
encounterCollection(conn, idResult)
}
\arguments{
\item{conn}{connection object}

\item{numeric}{Result number}
}
\value{
data.frame encounter results
}
\description{
From the result of the query i2b2, this function returns a list of visits matching.
}

