% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{exportData}
\alias{exportData}
\title{exportData}
\usage{
exportData(data, file)
}
\arguments{
\item{data}{Data type to export}

\item{file}{Name of CSV file}
}
\description{
Creates different CSV file
}

