% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getColumnFile}
\alias{getColumnFile}
\title{getColumnFile}
\usage{
getColumnFile(file)
}
\arguments{
\item{file}{Name of CSV file}
}
\value{
data.frame the columns
}
\description{
Reads a file from a directory
}

