% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getRole}
\alias{getRole}
\title{getRole}
\usage{
getRole(conn, idProject, idUser)
}
\arguments{
\item{conn}{connection object}

\item{character}{idProject Project ID}

\item{character}{idSession Session ID}
}
\value{
character the roles list
}
\description{
Determines the user's privileges. Columns to display depend on this privilege.
}

