% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getSetTableFromIdResult}
\alias{getSetTableFromIdResult}
\title{getSetTableFromIdResult}
\usage{
getSetTableFromIdResult(conn, idResult)
}
\arguments{
\item{conn}{connection object}

\item{numeric}{idResult Result ID}
}
\value{
character The tableset name
}
\description{
This function get the table name to ask for data extraction
}

