% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getValueFromConfFile}
\alias{getValueFromConfFile}
\title{getValueFromConfFile}
\usage{
getValueFromConfFile(file, pattern)
}
\arguments{
\item{character}{The read file path}

\item{character}{pattern The field to read}
}
\value{
character the value
}
\description{
Reading the connection file
}

