% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{package}
\name{i2b2rplugin}
\alias{PATH_CONFIG}
\alias{i2b2rplugin}
\alias{i2b2rplugin-package}
\title{i2b2rplugin}
\format{An object of class \code{character} of length 1.}
\usage{
PATH_CONFIG
}
\description{
Contains basic fonction to dialog with i2b2 database
and data model. Supposed to deal with :
- Export data feature (first release)
- Statistical visualisation (second release)
throught OpenCPU R webservice software calls
}
\keyword{datasets}

