% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{patientCollection}
\alias{patientCollection}
\title{patientCollection}
\usage{
patientCollection(conn, idResult)
}
\arguments{
\item{conn}{connection object}

\item{numeric}{Result number}
}
\value{
data.frame patient results
}
\description{
From the result of the query i2b2, this function returns a list of facts matching.
}

