% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{providerDimension}
\alias{providerDimension}
\title{providerDimension}
\usage{
providerDimension(conn)
}
\arguments{
\item{conn}{connection object}
}
\value{
data.frame results
}
\description{
From the result of the query i3b2, this function returns a list of matching patients.
}

