% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{verifActifUser}
\alias{verifActifUser}
\title{verifActifUser}
\usage{
verifActifUser(conn, idSession, idUser)
}
\arguments{
\item{conn}{connection object}

\item{character}{idSession Session ID}

\item{character}{idUser User ID}
}
\description{
Verifies that the user is active and authorized to export the database
}

