% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{verifActiveSession}
\alias{verifActiveSession}
\title{verifActiveSession}
\usage{
verifActiveSession(conn, idProject, idSession, idUser)
}
\arguments{
\item{conn}{connection object}

\item{character}{idProject Project ID}

\item{character}{idSession Session ID}

\item{character}{idUser User ID}
}
\description{
Verifies that the user's session is on
}

