% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{verifIdProject}
\alias{verifIdProject}
\title{verifIdProject}
\usage{
verifIdProject(conn, idProject)
}
\arguments{
\item{conn}{connection object}

\item{idProjet}{L'idProjet a verifier}
}
\description{
Checks if project ID is valid
}

