% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{verifIdResult}
\alias{verifIdResult}
\title{verifIdResult}
\usage{
verifIdResult(conn, idResult)
}
\arguments{
\item{conn}{connection object}

\item{idResult}{Result ID}
}
\description{
Checks if result ID is valid
}

