% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{verifQueryOwner}
\alias{verifQueryOwner}
\title{verifQueryOwner}
\usage{
verifQueryOwner(conn, idSession, idProject, idResult, idUser)
}
\arguments{
\item{conn}{connection object}

\item{character}{idSession Session ID}

\item{character}{idProject Project ID}

\item{numeric}{idResult Result ID}

\item{character}{idUser User ID}
}
\description{
This function ensures that the current user is the author of the petition.
}

